desktop/requirements.txt
# gRPC
grpcio==1.60.0
grpcio-tools==1.60.0
protobuf==4.25.1

# OpenAI
openai==1.10.0

# Anthropic Claude
anthropic==0.18.1

# Audio processing
pyaudio==0.2.14
pydub==0.25.1
numpy==1.26.3

# Configuration
python-dotenv==1.0.0

# Logging
colorlog==6.8.0

# HTTP client for OAuth
requests==2.31.0

# Optional: Local speech recognition
# SpeechRecognition==3.10.1
# whisper-openai==1.0.0

# Optional: Local TTS
# pyttsx3==2.90
# TTS==0.22.0  # Coqui TTS
